.travis.yml: Make the env variable a single value
It seems the official way of defining the travis env
is - ENV=VAR only and using the dictionary is a hacky
way.  Due to that fact, they use the white space as
a separater of the env variable.  Hense, I've just
overcome that short comings by defining the multiple
keys so that we can avoid the white space separation
inside the env variable.