utils/Dockerfile.*: Use keinohguchi's ansible-modules-core
This is due to the temp fix in ops_template.py, to handle
the non-string key returned by ops.dc.read().

I believe the actual fix should be in the ops.dc.read()
side but putting the change in the ops_template.py side
so that we can get going with the ansible role testing.