README.md: Change the seven switch topology
We now have seven switches in the topology, which
requires to change both docker-compose.yaml as well
as docker-network script.  With this change, we have
two leaf/ToR switches under each spine switches
and will help to simulate the VLAN based operation
without changing the topology.  This also make
the dc-on-docker similar to the dc-on-gns3, as it
has a limitation of the number of interfaces to
three, which prevents to create a six switch topology
we had before.

Here is the current topology, as shown in the README.md:

```

                          |1
                       +--+---+
                       | fab1 |
                       +-+--+-+
                         |2 |3
                    +----+  +----+
                    |1           |1
                +---+----+  +----+---+
                | spine1 |  | spine2 |
                +-+---+--+  +--+---+-+
                  |2  |3       |2  |3
           +------+  ++        ++  +------+
           |1        |1         |1        |1
       +---+---+ +---+---+  +---+---+ +---+---+
       | leaf1 | | leaf2 |  | leaf3 | | leaf4 |
       +-+---+-+ +-+---+-+  +-+---+-+ +-+---+-+
         |2  |3    |2  |3     |2  |3    |2  |3
```